task6fixa

+++ hello_world/6-size.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 /**
- *main - main
+ *main- main
  *Return: always to 0
 */
 int main(void)
@@ -12,11 +12,10 @@ long long int c;
 char d;
 float f;
 
-printf("Size of char: %lu byte(s)\n", sizeof(d));
-printf("Size an int: %lu byte(s)\n", sizeof(a));
-printf("Size of long: %lu byte(s)\n", sizeof(b));
-printf("Size of long long: %lu byte(s)\n", sizeof(c));
-printf("Size of float: %lu byte(s)\n", sizeof(f));
-
+printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(d));
+printf("Size an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of long: %lu byte(s)\n", (unsigned long)sizeof(b));
+printf("Size of long long: %lu byte(s)\n", (unsigned long)sizeof(c));
+printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
